refactor(tasks): move Go2 Arm manip-loco owner [skip ci]
