+++ scripts/manip_loco/diagnose_go2_arm_ik.py
@@ -16,7 +16,7 @@ if str(ROOT_DIR) not in sys.path:
 
 from unilab.base import registry
 from unilab.base.registry import ensure_registries
-from unilab.envs.locomotion.go2_arm.manip_loco import RewardConfig
+from unilab.tasks.locomotion.go2_arm.manip_loco import RewardConfig
 from unilab.utils.rotation import np_matrix_from_quat
