+++ scripts/manip_loco/play_go2_arm_ik_only.py
@@ -19,7 +19,7 @@ if str(ROOT_DIR) not in sys.path:
     sys.path.insert(0, str(ROOT_DIR))
 
 from unilab.envs.locomotion.go2_arm.base import build_go2_arm_position_gains
-from unilab.envs.locomotion.go2_arm.manip_loco import Go2ArmManipLocoCfg
+from unilab.tasks.locomotion.go2_arm.manip_loco import Go2ArmManipLocoCfg
 
 TARGET_BODY = "ik_mocap_target"
