+++ src/unilab/tasks/__init__.py
@@ -12,7 +12,7 @@ __unilab_registry_modules__ = (
     "unilab.tasks.locomotion.go2",
     "unilab.tasks.locomotion.go2w",
     "unilab.envs.locomotion.g1",
-    "unilab.envs.locomotion.go2_arm",
+    "unilab.tasks.locomotion.go2_arm",
     "unilab.tasks.locomotion.a2",
     "unilab.envs.manipulation.allegro_inhand",
     "unilab.envs.manipulation.sharpa_inhand",
