+++ src/unilab/tasks/locomotion/go2_arm/__init__.py
@@ -1,0 +1,3 @@
+from .manip_loco import Go2ArmManipLocoCfg, Go2ArmManipLocoEnv
+
+__all__ = ["Go2ArmManipLocoCfg", "Go2ArmManipLocoEnv"]
