+++ tests/tasks/test_package_boundary.py
@@ -16,7 +16,7 @@ _TASK_REGISTRY_MODULES = (
     "unilab.tasks.locomotion.go2",
     "unilab.tasks.locomotion.go2w",
     "unilab.envs.locomotion.g1",
-    "unilab.envs.locomotion.go2_arm",
+    "unilab.tasks.locomotion.go2_arm",
     "unilab.tasks.locomotion.a2",
     "unilab.envs.manipulation.allegro_inhand",
     "unilab.envs.manipulation.sharpa_inhand",
